+++ packages/helpers/src/types.ts
@@ -1,3 +1,10 @@
 export type PropertyName<TTarget> = keyof TTarget & string;
 
 export type BindingProperty<TTarget> = TTarget extends Map<infer K, any> ? K : PropertyName<TTarget>;
+
+export type PropertyType<TTarget extends Map<any, any> | {}, TProperty extends BindingProperty<TTarget>> = TTarget extends Map<
+  any,
+  infer V
+>
+  ? V
+  : TTarget[TProperty];