+++ packages/views/src/index.ts
@@ -1,9 +1,11 @@
 export * from "./bindingComponent";
+export * from "./bindingProps";
 export { default as createDataIdHandler } from "./dataIdHandler";
 export * from "./fieldHelpers";
 export { default as createMemoizedHandler } from "./memoizedDataHandler";
 export { default as preventDefault } from "./preventDefault";
 export * from "./types";
+export * from "./useBinding";
 export { default as useDisposable } from "./useDisposable";
 export { default as useScreenLifecycle } from "./useScreenLifecycle";
 export { default as View } from "./view";