+++ stories/src/controls/formField.tsx
@@ -13,9 +13,9 @@ export interface ChildProps {
 }
 
 export const FormField: React.FunctionComponent<IFormFieldProps<any, ChildProps> & FieldProps> = observer(props => {
-  const validationMessage = props.target && props.property && getValidationMessage(props.target, props.property);
-  const isDirty = getDirtyFlag(props.target, props.property);
-  const bordercolor = !!validationMessage ? "red" : (isDirty ? "green" : "black");
+  const validationMessage = props.target && props.property && getValidationMessage(props.target, props.property as any);
+  const isDirty = getDirtyFlag(props.target, props.property as any);
+  const bordercolor = !!validationMessage ? "red" : isDirty ? "green" : "black";
 
   const childProps: ChildProps = {
     bordercolor,
@@ -25,7 +25,8 @@ export const FormField: React.FunctionComponent<IFormFieldProps<any, ChildProps>
 
   return (
     <div style={{ border: `1px solid ${bordercolor}`, padding: 10 }}>
-      <label>{props.label}:</label><br />
+      <label>{props.label}:</label>
+      <br />
       {content}
       <div style={{ color: bordercolor }}>{validationMessage}</div>
       {isDirty && <div style={{ color: "green" }}>Needs save</div>}
@@ -33,10 +34,14 @@ export const FormField: React.FunctionComponent<IFormFieldProps<any, ChildProps>
   );
 });
 
-export function fieldForType<TTarget>(target: TTarget): React.FunctionComponent<IFormFieldProps<TTarget, ChildProps> & FieldProps> {
-  return FormField;
+export function fieldForType<TTarget>(
+  target: TTarget
+): React.FunctionComponent<IFormFieldProps<TTarget, ChildProps> & FieldProps> {
+  return FormField as any;
 }
 
-export function fieldForTarget<TTarget>(target: TTarget): React.FunctionComponent<IFormFieldProps<TTarget, ChildProps> & FieldProps> {
+export function fieldForTarget<TTarget>(
+  target: TTarget
+): React.FunctionComponent<IFormFieldProps<TTarget, ChildProps> & FieldProps> {
   return (props: any) => <FormField {...props} target={target} />;
 }